(finalizers): refactor
