+++ controllers/druid/handler.go
@@ -80,35 +80,12 @@ func deployDruidCluster(ctx context.Context, sdk client.Client, m *v1alpha1.Drui
 		return err
 	}
 
-	/*
-		Default Behavior: Finalizer shall be always executed resulting in deletion of pvc post deletion of Druid CR
-		When the object (druid CR) has for deletion time stamp set, execute the finalizer.
-		Finalizer shall execute the following flow :
-		1. Get sts List and PVC List
-		2. Range and Delete sts first and then delete pvc. PVC must be deleted after sts termination has been executed
-			else pvc finalizer shall block deletion since a pod/sts is referencing it.
-		3. Once delete is executed we block program and return.
-	*/
+	if m.GetDeletionTimestamp() != nil {
+		return executeFinalizers(ctx, sdk, m, emitEvents)
+	}
 
-	if m.Spec.DisablePVCDeletionFinalizer == false {
-		md := m.GetDeletionTimestamp() != nil
-		if md {
-			return executeFinalizers(ctx, sdk, m, emitEvents)
-		}
-		/*
-			If finalizer isn't present add it to object meta.
-			In case cr is already deleted do not call this function
-		*/
-		cr := checkIfCRExists(ctx, sdk, m, emitEvents)
-		if cr {
-			if !ContainsString(m.ObjectMeta.Finalizers, finalizerName) {
-				m.SetFinalizers(append(m.GetFinalizers(), finalizerName))
-				_, err := writers.Update(context.Background(), sdk, m, m, emitEvents)
-				if err != nil {
-					return err
-				}
-			}
-		}
+	if err := addFinalizers(ctx, sdk, m, emitEvents); err != nil {
+		return err
 	}
 
 	for _, elem := range allNodeSpecs {
@@ -572,61 +549,6 @@ func setPVCLabels(ctx context.Context, sdk client.Client, drd *v1alpha1.Druid, e
 		}
 	}
 	return nil
-}
-
-func executeFinalizers(ctx context.Context, sdk client.Client, m *v1alpha1.Druid, emitEvents EventEmitter) error {
-
-	if ContainsString(m.ObjectMeta.Finalizers, finalizerName) {
-		pvcLabels := map[string]string{
-			"druid_cr": m.Name,
-		}
-
-		pvcList, err := readers.List(ctx, sdk, m, pvcLabels, emitEvents, func() objectList { return &v1.PersistentVolumeClaimList{} }, func(listObj runtime.Object) []object {
-			items := listObj.(*v1.PersistentVolumeClaimList).Items
-			result := make([]object, len(items))
-			for i := 0; i < len(items); i++ {
-				result[i] = &items[i]
-			}
-			return result
-		})
-		if err != nil {
-			return err
-		}
-
-		stsList, err := readers.List(ctx, sdk, m, makeLabelsForDruid(m.Name), emitEvents, func() objectList { return &appsv1.StatefulSetList{} }, func(listObj runtime.Object) []object {
-			items := listObj.(*appsv1.StatefulSetList).Items
-			result := make([]object, len(items))
-			for i := 0; i < len(items); i++ {
-				result[i] = &items[i]
-			}
-			return result
-		})
-		if err != nil {
-			return err
-		}
-
-		msg := fmt.Sprintf("Trigerring finalizer for CR [%s] in namespace [%s]", m.Name, m.Namespace)
-		//		sendEvent(sdk, m, v1.EventTypeNormal, DruidFinalizer, msg)
-		logger.Info(msg)
-		if err := deleteSTSAndPVC(ctx, sdk, m, stsList, pvcList, emitEvents); err != nil {
-			return err
-		} else {
-			msg := fmt.Sprintf("Finalizer success for CR [%s] in namespace [%s]", m.Name, m.Namespace)
-			//			sendEvent(sdk, m, v1.EventTypeNormal, DruidFinalizerSuccess, msg)
-			logger.Info(msg)
-		}
-
-		// remove our finalizer from the list and update it.
-		m.ObjectMeta.Finalizers = RemoveString(m.ObjectMeta.Finalizers, finalizerName)
-
-		_, err = writers.Update(ctx, sdk, m, m, emitEvents)
-		if err != nil {
-			return err
-		}
-
-	}
-	return nil
-
 }
 
 func execCheckCrashStatus(ctx context.Context, sdk client.Client, nodeSpec *v1alpha1.DruidNodeSpec, m *v1alpha1.Druid, event EventEmitter) {
